+++ src/Origin/Properties/CompetenceManagement/CompetenceManagementProperties.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace srag\Plugins\Hub2\Origin\Properties\CompetenceManagement;
+
+use srag\Plugins\Hub2\Origin\Properties\OriginProperties;
+
+/**
+ * Class CompetenceManagementProperties
+ *
+ * @package srag\Plugins\Hub2\Origin\Properties\CompetenceManagement
+ *
+ * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
+ */
+class CompetenceManagementProperties extends OriginProperties implements ICompetenceManagementProperties {
+
+}