+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [x]
+- Competence management origin type
 - Fix install log
 - Fix language keys with line break
 - Fix read fallback parent ref id read